requirements.txt
alembic==1.11.3
aniso8601==9.0.1
asn1crypto==0.24.0
autopep8==2.0.4
bcrypt==4.0.1
blinker==1.6.2
botocore==1.12.189
cffi==1.15.1
click==8.1.7
cryptography==2.7
docutils==0.14
Flask==2.3.3
Flask-API==3.1
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Migrate==4.0.4
Flask-RESTful==0.3.10
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
gunicorn==19.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.9.4
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pycparser==2.19
PyJWT==2.8.0
python-dateutil==2.8.0
python-dotenv==1.0.0
python-editor==1.0.4
pytz==2023.3
s3transfer==0.2.1
six==1.16.0
SQLAlchemy==2.0.20
tomli==2.0.1
typing_extensions==4.7.1
urllib3==1.25.3
Werkzeug==2.3.7
